+++ src/main/java/dynamicquad/agilehub/global/header/ReasonDto.java
@@ -1,4 +1,4 @@
-package dynamicquad.agilehub.global.header.status;
+package dynamicquad.agilehub.global.header;
 
 import lombok.Builder;
 import lombok.Getter;
