+++ src/main/java/dynamicquad/agilehub/global/header/status/BaseStatus.java
@@ -1,5 +1,7 @@
 package dynamicquad.agilehub.global.header.status;
 
+import dynamicquad.agilehub.global.header.ReasonDto;
+
 public interface BaseStatus {
     public ReasonDto getReason();
